Fix double vote check. Change accumulation votings logic

+++ src/main/java/by/andd3dfx/voting/dto/response/VotingsResponse.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class VotingsResponse {
 
-    private Map<String, Long> votings;
+    private Map<String, Integer> votings;
 }
